model utils add ldlib plugin
