+++ settings.gradle
@@ -18,15 +18,15 @@ dependencyResolutionManagement {
 
         // Mod Dependencies Versions
         // Common
-        def jeiVersion = "15.2.0.27"
-        def reiVersion = "12.0.684"
-        def emiVersion = "1.1.4"
-        def ae2Version = "15.0.18"
+        def jeiVersion = "15.8.0.11"
+        def reiVersion = "12.1.725"
+        def emiVersion = "1.1.10"
+        def ae2Version = "15.2.11"
         def kjsVersion = "2001.6.4-build.120"
         def gtmVersion = "1.3.1"
 
         // Forge
-        def forgeVersion = "47.1.3"
+        def forgeVersion = "47.2.0"
         def registrateForgeVersion = "MC1.20-1.3.3"
         def jadeForgeVersion = "11.7.1"
         def worldStripperForgeFile = "4578579"
@@ -42,7 +42,7 @@ dependencyResolutionManagement {
         def configurationVersion = "2.2.0"
         def ldLibVersion = "1.0.26"
         def mixinextrasVersion = "0.2.0"
-        def shimmerVersion = "0.2.2"
+        def shimmerVersion = "0.2.4"
         def spotlessVersion = "7.0.0.BETA1"
 
         forge {
