+++ src/main/java/com/epimorphismmc/monomorphism/client/model/ModelFactory.java
@@ -8,16 +8,23 @@ import net.minecraft.client.renderer.texture.MissingTextureAtlasSprite;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.BlockModelRotation;
 import net.minecraft.client.resources.model.Material;
+import net.minecraft.client.resources.model.ModelBaker;
 import net.minecraft.client.resources.model.ModelBakery;
 import net.minecraft.client.resources.model.ModelManager;
 import net.minecraft.client.resources.model.ModelResourceLocation;
+import net.minecraft.client.resources.model.ModelState;
 import net.minecraft.client.resources.model.UnbakedModel;
+import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Mth;
 import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+
+import org.joml.Quaternionf;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,7 +32,7 @@ import java.util.List;
 import static com.epimorphismmc.monomorphism.client.utils.ClientUtils.blockRendererDispatcher;
 import static com.epimorphismmc.monomorphism.client.utils.ClientUtils.mc;
 
-public class ModelAssistant {
+public class ModelFactory {
 
     public static final ItemModelGenerator ITEM_MODEL_GENERATOR = new ItemModelGenerator();
 
@@ -47,6 +54,11 @@ public class ModelAssistant {
      */
     public static ModelBakery modeBakery() {
         return modelManager().getModelBakery();
+    }
+
+    // TODO
+    public static ModelBaker getModeBaker() {
+        return null;
     }
 
     public static UnbakedModel getUnbakedModel(ResourceLocation modelLocation) {
@@ -139,5 +151,57 @@ public class ModelAssistant {
 
     public static BakedModel getMissingModel() {
         return modelManager().getMissingModel();
+    }
+
+    public static Quaternionf getQuaternion(Direction facing) {
+        return switch (facing) {
+            case UP -> new Quaternionf().rotateXYZ(Mth.HALF_PI, 0, 0);
+            case DOWN -> new Quaternionf().rotateXYZ(-Mth.HALF_PI, 0, 0);
+            case EAST -> new Quaternionf().rotateXYZ(0, -Mth.HALF_PI, 0);
+            case WEST -> new Quaternionf().rotateXYZ(0, Mth.HALF_PI, 0);
+            case SOUTH -> new Quaternionf().rotateXYZ(0, Mth.PI, 0);
+            case NORTH -> new Quaternionf();
+        };
+    }
+
+    public static ModelState getRotation(Direction facing) {
+        return switch (facing) {
+            case DOWN -> BlockModelRotation.X90_Y0;
+            case UP -> BlockModelRotation.X270_Y0;
+            case NORTH -> BlockModelRotation.X0_Y0;
+            case SOUTH -> BlockModelRotation.X0_Y180;
+            case WEST -> BlockModelRotation.X0_Y270;
+            case EAST -> BlockModelRotation.X0_Y90;
+        };
+    }
+
+    public static Direction modelFacing(Direction side, Direction frontFacing) {
+        if (side == frontFacing) return Direction.NORTH;
+        if (frontFacing == Direction.NORTH) return side;
+        if (frontFacing == Direction.SOUTH) {
+            if (side.getAxis() == Direction.Axis.Y) return side;
+            return side.getOpposite();
+        }
+        if (frontFacing == Direction.EAST) {
+            if (side.getAxis() == Direction.Axis.Y) return side;
+            return side.getCounterClockWise();
+        }
+        if (frontFacing == Direction.WEST) {
+            if (side.getAxis() == Direction.Axis.Y) return side;
+            return side.getClockWise();
+        }
+        if (frontFacing == Direction.UP) {
+            if (side == Direction.DOWN) return Direction.SOUTH;
+            if (side.getAxis() == Direction.Axis.X) return side;
+            if (side == Direction.SOUTH) return Direction.UP;
+            if (side == Direction.NORTH) return Direction.DOWN;
+        }
+        if (frontFacing == Direction.DOWN) {
+            if (side == Direction.UP) return Direction.SOUTH;
+            if (side.getAxis() == Direction.Axis.X) return side;
+            if (side == Direction.SOUTH) return Direction.DOWN;
+            if (side == Direction.NORTH) return Direction.UP;
+        }
+        return side;
     }
 }
