+++ src/main/java/com/epimorphismmc/monomorphism/integration/ldlib/MOLDLibPlugin.java
@@ -1,0 +1,14 @@
+package com.epimorphismmc.monomorphism.integration.ldlib;
+
+import com.epimorphismmc.monomorphism.syncdata.SyncedFieldAccessors;
+
+import com.lowdragmc.lowdraglib.plugin.ILDLibPlugin;
+import com.lowdragmc.lowdraglib.plugin.LDLibPlugin;
+
+@LDLibPlugin
+public class MOLDLibPlugin implements ILDLibPlugin {
+    @Override
+    public void onLoad() {
+        SyncedFieldAccessors.init();
+    }
+}
